+++ src/main/java/cart/controller/ExceptionAdvice.java
@@ -1,0 +1,41 @@
+package cart.controller;
+
+import cart.dto.ErrorDto;
+import javax.validation.ConstraintViolationException;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@Slf4j
+@RestControllerAdvice
+public class ExceptionAdvice {
+
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    public ErrorDto handleException(MethodArgumentNotValidException exception) {
+        FieldError foundError = (FieldError) exception.getBindingResult()
+                .getAllErrors().stream()
+                .findFirst()
+                .orElseThrow(() -> new IllegalStateException("예외 메시지 정보가 존재하지 않습니다."));
+        log.error("MethodArgumentNotValid message={}", foundError.getDefaultMessage());
+        return new ErrorDto(foundError.getDefaultMessage());
+    }
+
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    @ExceptionHandler(ConstraintViolationException.class)
+    public ErrorDto handleException(ConstraintViolationException exception) {
+        log.error("ConstraintViolationException message={}", exception.getMessage());
+        return new ErrorDto(exception.getMessage());
+    }
+
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    @ExceptionHandler(IllegalArgumentException.class)
+    public ErrorDto handleException(IllegalArgumentException exception) {
+        log.error("IllegalArgumentException message={}", exception.getMessage());
+        return new ErrorDto(exception.getMessage());
+    }
+}
